Update from dherring/able, PR#1, credit stacksmith
Per stacksmith: https://github.com/stacksmith

Modified (prompt...) to show the current package, as a nickname if po…

…ssible, unless a specific prompt is requested.

Also, changed the output stream to force the buffer to output if there is a delay of 1/5 of a second.  This makes quickload and other functions that take time and print dots on the screen work.
The combined effect of these small changes is that the user experience is much closer to Slime.